backend/pyproject.toml
[project]
name = "elevora"
version = "0.1.0"
description = ""
authors = [
    {name = "Kashan Haider",email = "hkashan.dev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.32.3,<3.0.0",
    "beautifulsoup4>=4.13.4,<5.0.0",
    "fastapi>=0.115.12,<0.116.0",
    "uvicorn>=0.34.2,<0.35.0",
    "matplotlib>=3.10.1,<4.0.0",
    "pandas>=2.2.3,<3.0.0",
    "seaborn>=0.13.2,<0.14.0",
    "selenium>=4.32.0,<5.0.0",
    "sqlalchemy>=2.0.41,<3.0.0",
    "psycopg2-binary>=2.9.10,<3.0.0",
    "alembic>=1.15.2,<2.0.0",
    "pydantic[email]>=2.11.4,<3.0.0",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (<4.1.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false